Fix lifetime issues on iter() and iter_mut()
Changing from a custom Iter implementation to simply using the
core::slice::Iter/IterMut implementations means that the iter() and
iter_mut() functions no longer need to be tied to the lifetime of the
FixedVec. In the case of iter_mut(), keeping the lifetime will actually
cause ownership errors because the iterator will "attach" itself to the
lifetime of the FixedVec instead of the lifetime of the mutable borrow.

Signed-off-by: Nick Stevens <nick@bitcurry.com>