Update for latest nightly, remove example for now
Although #![no_std] has been stabilized, the API for handling programs
with a main function that use no_std has not been stabilized. As a
result, the existing example passes on OSX but fails on Linux, for
reasons unrelated to the code.

Long story short, generating no_std executables is most definitely still
an unstable feature.

Signed-off-by: Nick Stevens <nick@bitcurry.com>