doc: fix compile/link error of main example
The main example (which links using nostd) was failing due to the
missing _Unwind_Resume symbol. This symbol is the landing pad for
unwinding panics, which usually wouldn't be used in embedded code. For
embedded executables, the best approach is the add the `panic = "abort"`
profile option to remove these landing pads.

However, in the doctest case, it is important that the stack still
unwind if there's a failure - the entire doctest framework shouldn't
abort because one doctest failed. So setting `panic = "abort"` is not
ideal.

Since the main example doesn't actually run, it is possible to "fake"
the landing pads by just creating a stub function that loops forever.
This way other doc tests still unwind on failure, and the main example
compiles and links cleanly.

In addition to creating the stub function, this commit also adds a blurb
to the documentation recommending that users of the library investigate
profiles and `panic = "abort"` settings for their own applications.

Signed-off-by: Nick Stevens <nick@bitcurry.com>