Support and test for "libcore" stable
As of Rust 1.6, libcore (i.e. #![no_std]) is stable. As a result,
there's no reason that fixedvec cannot compile and test against stable
versions of Rust.

This commit updates the CI scripts to test 1.6-1.11, beta, and nightly.
It also updates the example to remove unstable code, and updates the
README to make explicit that the library does build on stable.

Also add to the README explicit mention that no_std binaries are _not_
yet stable and will still require nightly Rust.

Signed-off-by: Nick Stevens <nick@bitcurry.com>